+++ packages/workspace-fusion/src/lib/integrations/power-bi/utils/addFusionPowerBi.tsx
@@ -1,7 +1,7 @@
 import { WorkspaceViewController } from '@equinor/workspace-react';
 import { FusionPowerBiConfigurator } from '../../../classes';
 import { FusionMediator, FusionWorkspaceError, WorkspaceTabNames } from '../../../types';
-import { FusionPowerBiConfig } from '..';
+import { FusionPowerBiConfig } from '../';
 import { addPowerBi } from './addPowerBi';
 
 export function addFusionPowerBi<
